refactor(input box): seperate InputBox, DropdownInputBox, ChipInputBox
